# openapi.yml
openapi: 3.0.3
info:
  title: HMS Backend API
  version: 1.0.0
  description: |
    OpenAPI specification based solely on the application's actual Express routes and validation rules.
  contact:
    name: HMS Admin
    email: mrfortrancer@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://hms.fortrancer.in
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or missing required information.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: A resource conflict occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong.
        statusCode:
          type: integer
          example: 500
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        details:
          type: object
          additionalProperties: true
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
...
    SetPasswordRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: NewStrongPass123!
...
    PatientSignUpRequest:
      type: object
      required:
        - name
        - role
        - email
        - gender
        - phone
        - dob
        - address
        - bloodGroup
      properties:
        name:
          type: string
          example: Rahul Sharma
        role:
          type: string
          example: Patient
        email:
          type: string
          format: email
          example: rahul.sharma@example.com
        password:
          type: string
          example: PatientPass123
        status:
          type: string
          example: Active
        phone:
          type: string
          example: +91-9876543210
        gender:
          type: string
          example: Male
        dob:
          type: string
          format: date
          example: 1992-05-14
        address:
          type: string
          example: 12, MG Road, Bengaluru
        bloodGroup:
          type: string
          enum:
            - A+
            - A-
            - B+
            - B-
            - AB+
            - AB-
            - O+
            - O-
          example: O+
        allergies:
          type: string
          example: Peanuts
        emergencyContact:
          type: string
          example: +91-9998887776
    ResetPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
...
    PermissionsResponse:
      type: object
      additionalProperties: true
paths:
...
  /auth/set-password:
    post:
      tags: [Auth]
      summary: Set password for first-time user
      description: Sets a password for a newly created user.
      operationId: setPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      responses:
        '200':
          description: Password updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/patientSignUp:
    post:
      tags: [Auth]
      summary: Register patient
      description: Creates a patient profile and linked user account.
      operationId: patientSignUp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientSignUpRequest'
      responses:
        '201':
          description: Patient account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/getPermissions:
    get:
      tags: [Auth]
      summary: Get role permissions
      description: Retrieves permissions for the supplied role.
      operationId: getPermissions
      security: []
      parameters:
        - in: query
          name: role
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permissions returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/refresh-token:
    get:
      tags: [Auth]
      summary: Refresh access token
      description: Issues a fresh access token using the refresh token cookie.
      operationId: refreshToken
      security: []
      responses:
        '200':
          description: Access token refreshed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Sends password reset email if the account exists.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Reset link sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/verify-reset-password:
    get:
      tags: [Auth]
      summary: Verify reset password token
      description: Validates reset token and enables password reset.
      operationId: verifyResetPassword
      security: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: reset_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reset token verified successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
...